pyproject.toml
[tool.poetry]
name = "hitster"
version = "0.1.0"
description = ""
authors = ["fjlein <fjlein@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
typst = "^0.13.0"
spotipy = "^2.25.1"
qrcode = "^8.0"
pillow = "^11.1.0"
matplotlib = "^3.10.1"
pyjwt = "^2.8.0"
cryptography = "^41.0.0"
requests = "^2.31.0"
python-dotenv = "^1.1.1"
pydantic = "^2.12.3"
rich = "^13.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
